Rename dialog for a generic name
